Rename and refactor preprocess models file
Renamed `preprocess/models.py` to `preprocessing/models.py` for filename clarity. Consolidated classes and removed unused code to streamline preprocessing configurations and models.
